+++ src/__flowtests__/lints/sketchy-number.js
@@ -4,7 +4,7 @@ const React = { createElement: () => {} };
 
 const showBr: number = 42;
 
-export function Component() {
+export function Component(): $FlowFixMe {
   // $FlowExpectedError: sketchy-number-and
   return <div>{showBr && <br />}</div>;
 }